refactor: validateEmail and validatePassword into one file
