--- src/utils/validateEmail.ts
@@ -1,3 +1,0 @@
-export function validateEmail(email: string): boolean {
-  return /^[^\s@]+@[^\s@]+\.[^\s@]+$/.test(email) && !email.includes('/')
-}
